requirements.txt
psycopg2-binary~=2.8
pytest~=5.4
pytest-cov~=2.10
